Move booting to separate file
